+++ mygpiod/actions/http.c
@@ -18,23 +18,87 @@
 #include <unistd.h>
 
 // private definitions
+static bool fork_call_curl(const char *method, const char *uri, const char *content_type, const char *postdata);
+static bool call_curl(const char *method, const char *uri, const char *content_type, const char *postdata);
 static size_t catch_output(void *ptr, size_t size, size_t nmemb, sds *output);
-static void run_http(const char *cmd);
 
 // public functions
 
 /**
- * Makes a http call in a new process
- * @param cmd command to parse
+ * Makes a http call in a new process. Parses the cmd.
+ * @param cmd command and it's options, format:
+ *            {GET|POST} {uri} [{content-type} {postdata}]
  * @returns true on success, else false
  */
 bool action_http(const char *cmd) {
+    int count = 0;
+    sds *args = sdssplitargs(cmd, &count);
+    if (count < 2) {
+        MYGPIOD_LOG_ERROR("Invalid number of arguments");
+        sdsfreesplitres(args, count);
+        return false;
+    }
+    bool rc = false;
+    if (strcasecmp(args[0], "post") == 0) {
+        if (count != 4) {
+            MYGPIOD_LOG_ERROR("Invalid number of arguments");
+            sdsfreesplitres(args, count);
+            return false;
+        }
+        rc = fork_call_curl(args[0], args[1], args[2], args[3]);
+    }
+    else {
+        rc = fork_call_curl(args[0], args[1], NULL, NULL);
+    }
+    sdsfreesplitres(args, count);
+    return rc;
+}
+
+/**
+ * Makes a http call in a new process
+ * @param method HTTP method: POST or GET
+ * @param uri HTTP Uri
+ * @param content_type Content-Type or NULL
+ * @param postdata data to post or NULL
+ * @returns true on success, else false
+ */
+bool action_http2(const char *method, const char *uri, const char *content_type, const char *postdata) {
+    bool rc = false;
+    if (strcasecmp(method, "post") == 0) {
+        if (content_type == NULL ||
+            postdata == NULL)
+        {
+            MYGPIOD_LOG_ERROR("Invalid number of arguments");
+            return false;
+        }
+        rc = fork_call_curl(method, uri, content_type, postdata);
+    }
+    else {
+        rc = fork_call_curl(method, uri, NULL, NULL);
+    }
+    return rc;
+}
+
+// private functions
+
+/**
+ * Forks and makes the http call
+ * @param method HTTP method: POST or GET
+ * @param uri HTTP Uri
+ * @param content_type Content-Type or NULL
+ * @param postdata data to post or NULL
+ * @return true on success, else false
+ */
+static bool fork_call_curl(const char *method, const char *uri, const char *content_type, const char *postdata) {
     errno = 0;
     int pid = fork();
     if (pid == 0) {
         // this is the child process
-        run_http(cmd);
-        exit(EXIT_SUCCESS);
+        bool rc = call_curl(method, uri, content_type, postdata);
+        if (rc == true) {
+            exit(EXIT_SUCCESS);
+        }
+        exit(EXIT_FAILURE);
     }
     else {
         if (pid == -1) {
@@ -46,67 +110,51 @@ bool action_http(const char *cmd) {
     }
 }
 
-// private functions
-
 /**
- * Connects to MPD and runs the command
- * @param cmd command and it's options, format:
- *            {GET|POST} {uri} [{content-type} {postdata}]
+ * Makes an HTTP call
+ * @param method HTTP method: POST or GET
+ * @param uri HTTP Uri
+ * @param content_type Content-Type or NULL
+ * @param postdata data to post or NULL
  * @return true on success, else false
  */
-static void run_http(const char *cmd) {
-    int count = 0;
-    sds *args = sdssplitargs(cmd, &count);
-    if (count < 2) {
-        MYGPIOD_LOG_ERROR("Invalid number of arguments");
-        sdsfreesplitres(args, count);
-        exit(EXIT_FAILURE);
-    }
-    bool has_body = false;
-    if (strcasecmp(args[0], "post") == 0) {
-        has_body = true;
-        if (count != 4) {
-            MYGPIOD_LOG_ERROR("Invalid number of arguments");
-            sdsfreesplitres(args, count);
-            exit(EXIT_FAILURE);
-        }
-    }
-
+static bool call_curl(const char *method, const char *uri, const char *content_type, const char *postdata) {
+    (void) method;
     curl_global_init(CURL_GLOBAL_ALL);
     CURL *curl = curl_easy_init();
     if (curl == NULL) {
         curl_global_cleanup();
-        sdsfreesplitres(args, count);
-        exit(EXIT_FAILURE);
+        return false;
     }
     sds resp_header = sdsempty();
     sds resp_body = sdsempty();
-    curl_easy_setopt(curl, CURLOPT_URL, args[1]);
+    sds data = sdsnew(postdata);
+    curl_easy_setopt(curl, CURLOPT_URL, uri);
     curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, (curl_write_callback)catch_output);
     curl_easy_setopt(curl, CURLOPT_WRITEDATA, &resp_body);
     curl_easy_setopt(curl, CURLOPT_HEADERDATA, &resp_header);
     struct curl_slist *slist = NULL;
-    if (has_body == true) {
-        sds header = sdscatfmt(sdsempty(), "Content-type: %s", args[2]);
+    if (postdata != NULL) {
+        sds header = sdscatfmt(sdsempty(), "Content-type: %s", content_type);
         slist = curl_slist_append(slist, header);
         if (slist == NULL) {
-            exit(EXIT_FAILURE);
+            return false;
         }
         sdsfree(header);
         curl_easy_setopt(curl, CURLOPT_HTTPHEADER, slist);
-        if (strncmp(args[3], "<</", 3) == 0) {
+        if (strncmp(postdata, "<</", 3) == 0) {
             // read file
-            sds file_path = sdsdup(args[3]);
+            sds file_path = sdsnew(postdata);
             sdsrange(file_path, 2, -1);
-            sdsclear(args[3]);
+            sdsclear(data);
             int nread;
-            args[3] = sds_getfile(args[3], file_path, &nread);
+            data = sds_getfile(data, file_path, &nread);
             sdsfree(file_path);
             if (nread == -1) {
-                exit(EXIT_FAILURE);
+                return false;
             }
         }
-        curl_easy_setopt(curl, CURLOPT_POSTFIELDS, args[3]);
+        curl_easy_setopt(curl, CURLOPT_POSTFIELDS, data);
     }
     char err_buf[CURL_ERROR_SIZE];
     curl_easy_setopt(curl, CURLOPT_ERRORBUFFER, err_buf);
@@ -134,13 +182,14 @@ static void run_http(const char *cmd) {
             MYGPIOD_LOG_DEBUG("%s", resp_body);
         }
     }
+    curl_global_cleanup();
     sdsfree(resp_header);
     sdsfree(resp_body);
-    curl_global_cleanup();
-    sdsfreesplitres(args, count);
+    sdsfree(data);
     if (res != CURLE_OK) {
-        exit(EXIT_FAILURE);
+        return false;
     }
+    return true;
 }
 
 /**