+++ mygpiod/actions/http.h
@@ -10,5 +10,6 @@
 #include <stdbool.h>
 
 bool action_http(const char *cmd);
+bool action_http2(const char *method, const char *uri, const char *content_type, const char *postdata);
 
 #endif