+++ mygpiod/lib/action.h
@@ -27,6 +27,9 @@ enum mygpiod_actions {
         MYGPIOD_ACTION_HTTP,      //!< Makes an HTTP call
         MYGPIOD_ACTION_MYMPD,     //!< Executes a myMPD script
     #endif
+    #ifdef MYGPIOD_ENABLE_ACTION_LUA
+        MYGPIOD_ACTION_LUA,       //!< Executes a lua function
+    #endif
 };
 
 /**